refactored to pypet2bids
