+++ .github/workflows/setup_and_cli_test_posix.yaml
@@ -14,10 +14,10 @@ jobs:
       ECAT_TEST_FOLDER: "cimbi36"
       REAL_TEST_ECAT_PATH: cimbi36/Gris_102_19_2skan-2019.04.30.13.04.41_em_3d.v
       SMALLER_ECAT_PATH: ${{ github.workspace }}/PETdata_in/Siemens_ecat/ECAT7_multiframe.v
-      TEST_ECAT_PATH: ${{ github.workspace }}/pettobids/tests/ECAT7_multiframe.v
-      OUTPUT_NIFTI_PATH: ${{ github.workspace}}/pettobids/tests/ECAT7_multiframe.nii
-      READ_ECAT_SAVE_AS_MATLAB: ${{ github.workspace }}/pettobids/tests/ECAT7_multiframe.mat
-      NIBABEL_READ_ECAT_SAVE_AS_MATLAB: ${{ github.workspace }}/pettobids/tests/ECAT7_multiframe.nibabel.mat
+      TEST_ECAT_PATH: ${{ github.workspace }}/pypet2bids/tests/ECAT7_multiframe.v
+      OUTPUT_NIFTI_PATH: ${{ github.workspace}}/pypet2bids/tests/ECAT7_multiframe.nii
+      READ_ECAT_SAVE_AS_MATLAB: ${{ github.workspace }}/pypet2bids/tests/ECAT7_multiframe.mat
+      NIBABEL_READ_ECAT_SAVE_AS_MATLAB: ${{ github.workspace }}/pypet2bids/tests/ECAT7_multiframe.nibabel.mat
       MATLAB_CODE_PATH: ${{ github.workspace }}/matlab/
 
     strategy:
@@ -50,7 +50,7 @@ jobs:
       - name: Install dependencies
         if: steps.cached-poetry-dependencies.outputs.cache-hit != 'true'
         run: |
-          cd pettobids
+          cd pypet2bids
           poetry install --no-interaction --no-root
 
       - name: Collect Ecat Dataset w/ curl
@@ -61,22 +61,22 @@ jobs:
 
       - name: Test CLI --help
         run: |
-          cd pettobids/
+          cd pypet2bids/
           source .venv/bin/activate
-          python -m pettobids.cli --help
+          python -m pypet2bids.cli --help
 
       - name: Test CLI Ecat Dump
         run: |
-          cd pettobids/
+          cd pypet2bids/
           source .venv/bin/activate
-          python -m pettobids.cli ../${{ env.REAL_TEST_ECAT_PATH }} --dump
+          python -m pypet2bids.cli ../${{ env.REAL_TEST_ECAT_PATH }} --dump
 
       - name: Copy Test File(s) to tests/
-        run: cp ${{ env.SMALLER_ECAT_PATH }} pettobids/tests/
+        run: cp ${{ env.SMALLER_ECAT_PATH }} pypet2bids/tests/
 
       - name: Run All Other Python Tests w/ Pytest
         run: |
-          cd pettobids
+          cd pypet2bids
           source .venv/bin/activate
           pytest tests/
