+++ pypet2bids/.gitignore
@@ -85,7 +85,7 @@ celerybeat-schedule
 
 # virtualenv
 .venv
-../pettobidsblah/venv/
+../pypet2bidsblah/venv/
 ENV/
 
 # Spyder project settings
