+++ pypet2bids/pypet2bids/cli.py
@@ -3,8 +3,8 @@ import os
 import pathlib
 import sys
 from os.path import join
-from pettobids.ecat import Ecat
-from pettobids.helper_functions import load_vars_from_config
+from pypet2bids.ecat import Ecat
+from pypet2bids.helper_functions import load_vars_from_config
 
 """
 simple command line tool to extract header and pixel information from ecat files and convert ecat to nifti.
