+++ pypet2bids/pypet2bids/ecat2nii.py
@@ -3,7 +3,7 @@ import datetime
 import nibabel
 import numpy
 import pathlib
-from pettobids.read_ecat import read_ecat
+from pypet2bids.read_ecat import read_ecat
 import os
 import pickle
