+++ pypet2bids/pypet2bids/template.env
