+++ pypet2bids/template_env/SiemensHRRTparameters.env
