+++ pypet2bids/tests/scannerparams.txt
